Added curse; health changed to float instead of int

+++ Semester6_Game/Assets/Scripts/HUD Canvas/HealthBar.cs
@@ -5,7 +5,7 @@ using UnityEngine.UI;
 
 public class HealthBar : MonoBehaviour {
 
-    private int maxHealth;
+    private float maxHealth;
     private float currentHealth;
     private Image healthFill;
